Seeds/Wateringcan/Pot Interactions
The User can now walk up to the watering can "Pick it up" and water the plant

The User can now walk up to the Seeds "Pick it Up" and plant a seed.

The User can leave and come back and if they watered the plant then a flower will slowly grow.